compute/lecture: Update unikraft demo
This commit updates the unikraft demo for cooperative scheduling by:

- Removing dependencies on the `kraft` cli tool
- Merging all functionalities in one script
- Updating `main.c` to use pthreads function for better readability

To use the script, use:
$ ./do.sh setup
$ ./do.sh build
$ ./do.sh run_coop
$ ./do.sh run_nocoop
