--- content/chapters/compute/lecture/demo/cooperative-scheduling/app-schedule-threads/kraft.yaml
@@ -1,14 +1,0 @@
----
-specification: '0.5'
-name: app-schedule-threads
-unikraft:
-  version: 6a1e028
-  kconfig:
-    - CONFIG_LIBUKSCHEDCOOP=y
-    - CONFIG_LIBUKSIGNAL=y
-targets:
-  - architecture: x86_64
-    platform: linuxu
-libraries:
-  newlib:
-    version: '6375324'
