--- content/chapters/compute/lecture/demo/cooperative-scheduling/cleanup.sh
@@ -1,5 +1,0 @@
-#! /bin/bash
-
-APP_NAME="app-schedule-threads"
-
-cd ${APP_NAME}/ && kraft clean -d
