--- content/chapters/compute/lecture/demo/cooperative-scheduling/run.sh
@@ -1,30 +1,0 @@
-#! /bin/bash
-
-APP_NAME="app-schedule-threads"
-
-if [[ -gt $# 1 ]]; then
-    echo "Usage: $0 [--cooperate]"
-    exit 1
-fi
-
-if [[ $# == 1 ]]; then
-    COOP=$1
-else 
-    COOP=""
-fi
-
-cd ${APP_NAME} || exit 2
-
-kraft init -F
-
-if [[ ${COOP} == "--cooperate" ]]; then
-    echo "Running cooperative scheduling"
-    kraft configure -F -y APPSCHEDULETHREADS_COOPERATE
-else
-    echo "Running non-cooperative scheduling"
-    kraft configure -F -n APPSCHEDULETHREADS_COOPERATE
-fi
-
-kraft build
-
-kraft run
