--- content/chapters/compute/lecture/demo/cooperative-scheduling/setup.sh
@@ -1,8 +1,0 @@
-#! /bin/bash
-
-sudo apt install -y --no-install-recommends build-essential libncurses-dev \
-    libyaml-dev flex git wget socat bison unzip uuid-runtime python3-pip python3
-
-pip3 install git+https://github.com/unikraft/kraft.git@staging
-
-kraft list update
