+++ content/chapters/compute/lecture/demo/cooperative-scheduling/workdir/apps/app-schedule-threads/.gitignore
