+++ content/chapters/compute/lecture/demo/cooperative-scheduling/workdir/apps/app-schedule-threads/Makefile
@@ -1,0 +1,10 @@
+UK_ROOT ?= $(PWD)/../../unikraft
+UK_LIBS ?= $(PWD)/../../libs
+LIBS = $(UK_LIBS)/musl
+
+all:
+	@$(MAKE) -C $(UK_ROOT) A=$(PWD) L=$(LIBS)
+
+$(MAKECMDGOALS):
+	@$(MAKE) -C $(UK_ROOT) A=$(PWD) L=$(LIBS) $(MAKECMDGOALS)
+
